requirements.txt
Django==3.0.3
djangorestframework_simplejwt
djangorestframework-xml
mock==4.0.3
ipdb
drf-yasg
